+++ CHANGES.rst
@@ -13,8 +13,12 @@ Features:
 * #83 Drop Python-3.6 support.
 * #127 Drop Python-3.7 support.
 * #83 Drop Django-2.2 support.
+* #134 Support adding COLUMN with UNIQUE; adding column without UNIQUE then add UNIQUE CONSTRAINT.
 
 Bug Fixes:
+
+* #134 inspectdb should suppress output 'id = AutoField(primary_key=True)'
+* #134 fix for decreasing size of column with default by create-copy-drop-rename strategy.
 
 3.0.0 (2022/02/27)
 ------------------